+++ bas_metadata_library/standards/iso_19115_common/data_identification_elements.py
@@ -51,6 +51,15 @@ class DataIdentification(MetadataRecordElement):
         if _credit != "":
             _["credit"] = _credit
 
+        status = Status(
+            record=self.record,
+            attributes=self.attributes,
+            xpath=f"{self.xpath}/gmd:identificationInfo/gmd:MD_DataIdentification/gmd:status",
+        )
+        _status = status.make_config()
+        if _status != "":
+            _["status"] = _status
+
         _contacts = []
         contacts_length = int(
             self.record.xpath(
@@ -272,6 +281,15 @@ class DataIdentification(MetadataRecordElement):
             )
             spatial_representation_type.make_element()
 
+        if "status" in self.attributes["identification"]:
+            status = Status(
+                record=self.record,
+                attributes=self.attributes,
+                parent_element=data_identification_element,
+                element_attributes=self.attributes["identification"],
+            )
+            status.make_element()
+
         if "spatial_resolution" in self.attributes["identification"]:  # pragma: no cover
             spatial_resolution = SpatialResolution(
                 record=self.record,
@@ -357,6 +375,40 @@ class Credit(MetadataRecordElement):
             credit_element = SubElement(self.parent_element, f"{{{self.ns.gmd}}}credit")
             credit_value = SubElement(credit_element, f"{{{self.ns.gco}}}CharacterString")
             credit_value.text = self.element_attributes["credit"]
+
+
+class Status(CodeListElement):
+    def __init__(
+        self,
+        record: MetadataRecord,
+        attributes: dict,
+        parent_element: Element = None,
+        element_attributes: dict = None,
+        xpath: str = None,
+    ):
+        super().__init__(
+            record=record,
+            attributes=attributes,
+            parent_element=parent_element,
+            element_attributes=element_attributes,
+            xpath=f"{xpath}/gmd:MD_ProgressCode",
+        )
+        self.code_list_values = [
+            "completed",
+            "historicalArchive",
+            "obsolete",
+            "onGoing",
+            "planned",
+            "required",
+            "underDevelopment",
+        ]
+        self.code_list = (
+            "http://standards.iso.org/ittf/PubliclyAvailableStandards/ISO_19139_Schemas/resources/"
+            "codelist/gmxCodelists.xml#MD_ProgressCode"
+        )
+        self.element = f"{{{self.ns.gmd}}}status"
+        self.element_code = f"{{{self.ns.gmd}}}MD_ProgressCode"
+        self.attribute = "status"
 
 
 class PointOfContact(MetadataRecordElement):