+++ tests/bas_metadata_library/test_standard_iso_19115_1.py
@@ -377,6 +377,26 @@ def test_identification_credit(get_record_response, config_name):
     assert identification_credit_value is True
 
 
+@pytest.mark.usefixtures("get_record_response")
+@pytest.mark.parametrize("config_name", list(configs_safe_v2.keys()))
+def test_identification_status(get_record_response, config_name):
+    record = get_record_response(standard=standard, config=config_name)
+    config = configs_safe_v2[config_name]
+
+    if "identification" not in config or "status" not in config["identification"]:
+        pytest.skip("record does not contain an identification status")
+
+    status_elements = record.xpath(
+        f"/gmd:MD_Metadata/gmd:identificationInfo/gmd:MD_DataIdentification/gmd:status/"
+        f"gmd:MD_ProgressCode[@codeList = 'http://standards.iso.org/ittf/PubliclyAvailableStandards/"
+        f"ISO_19139_Schemas/resources/codelist/gmxCodelists.xml#MD_ProgressCode' and @codeListValue = "
+        f"'{config['identification']['status']}']/text()  = "
+        f"'{config['identification']['status']}'",
+        namespaces=namespaces.nsmap(),
+    )
+    assert status_elements is True
+
+
 # noinspection PyUnboundLocalVariable
 def _resolve_points_of_contact_xpaths(point_of_contact_type, config):
     pocs = []